Refs #460 unit test for _to_md
